Create generate.py
File to generate two versions of oriented graphs. 



We input the minimum degree n of the candidate oriented graph. Then the function creates a node and n nodes that are connected to this node. 

These n nodes are all connected to each other, but that will not be enough nodes to satisfy the minimum degree requirements, so they have exterior degree of 1. 

The process continues with a new set of n - 1 nodes that are all connected and have exterior degree 2. 

The nodes with back arcs is a bit more complex because while the previous arcs have all been added in the direction of v_0 (the minimum degree node) towards the other nodes. This function adds an arc back to that node v_0. There is a system of inequalities that must hold for an oriented graph to hold and Seymour's conjecture to fail. 

The gap between the system of inequalities grows larger and larger as the graphs get bigger which allows for the presence of more back arcs. 

The question becomes how does this impact the problem space?